requirements.txt
fastapi
uvicorn
httpx[http2]
pycryptodome
PyJWT
protobuf
urllib3